API/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymssql
pyodbc
